+++ aiof.api.data/IAsset.cs
@@ -11,7 +11,7 @@ namespace aiof.api.data
         string Name { get; set; }
         string TypeName { get; set; }
         AssetType Type { get; set; }
-        float Value { get; set; }
+        decimal Value { get; set; }
         int? FinanceId { get; set; }
     }
 }
